# gh-118761: improve import time of ``pickle``