--- a/Lib/pickle.py
+++ b/Lib/pickle.py
@@ -31,7 +31,6 @@ from functools import partial
 import sys
 from sys import maxsize
 from struct import pack, unpack
-import re
 import io
 import codecs
 import _compat_pickle
@@ -188,8 +187,10 @@ BYTEARRAY8       = b'\x96'  # push bytearray
 NEXT_BUFFER      = b'\x97'  # push next out-of-band buffer
 READONLY_BUFFER  = b'\x98'  # make top of stack readonly
 
-__all__.extend([x for x in dir() if re.match("[A-Z][A-Z0-9_]+$", x)])
-
+__all__.extend([
+    x for x in dir()
+    if x.isupper() and x.isidentifier() and not x.startswith('_')
+])
 
 class _Framer: