--- a/Lib/pickle.py
+++ b/Lib/pickle.py
@@ -187,10 +187,7 @@ BYTEARRAY8       = b'\x96'  # push bytearray
 NEXT_BUFFER      = b'\x97'  # push next out-of-band buffer
 READONLY_BUFFER  = b'\x98'  # make top of stack readonly
 
-__all__.extend([
-    x for x in dir()
-    if x.isupper() and x.isidentifier() and not x.startswith('_')
-])
+__all__.extend(x for x in dir() if x.isupper() and not x.startswith('_'))
 
 class _Framer: