--- a/Misc/NEWS.d/next/Library/2025-01-10-13-34-33.gh-issue-118761.qRB8nS.rst
+++ b/Misc/NEWS.d/next/Library/2025-01-10-13-34-33.gh-issue-118761.qRB8nS.rst
@@ -1,0 +1,2 @@
+Improve import time of :mod:`pickle` by 25% by removing an un-necessary
+import to :mod:`re`. Patch by Bénédikt Tran.