--- a/Misc/NEWS.d/next/Library/2025-01-10-13-34-33.gh-issue-118761.qRB8nS.rst
+++ b/Misc/NEWS.d/next/Library/2025-01-10-13-34-33.gh-issue-118761.qRB8nS.rst
@@ -1,2 +1,3 @@
-Improve import time of :mod:`pickle` by 25% by removing an un-necessary
-import to :mod:`re`. Patch by Bénédikt Tran.
+Improve import time of :mod:`pickle` by 25% by removing an unnecessary
+regular expression. As such, :mod:`re` is no more implicitly available
+as ``pickle.re``. Patch by Bénédikt Tran.