We can remove the ``re`` import which takes quite a long time. Benchmarks were performed on a RELEASE build (no PGO, no LTO).

### PR

```sh
$ hyperfine --warmup 8 "./python -c 'import pickle'"
Benchmark 1: ./python -c 'import pickle'
  Time (mean ± σ):       7.8 ms ±   0.3 ms    [User: 6.6 ms, System: 1.3 ms]
  Range (min … max):     7.4 ms …   9.8 ms    337 runs
```

### Main

```sh
$ hyperfine --warmup 8 "./python -c 'import pickle'"
Benchmark 1: ./python -c 'import pickle'
  Time (mean ± σ):       9.7 ms ±   0.3 ms    [User: 8.3 ms, System: 1.4 ms]
  Range (min … max):     9.3 ms …  12.6 ms    282 runs
```

Since something that is no more present in the global namespace is removed, I've added a NEWS entry and a detailed changelog.