@picnixz 2025-01-12
I'll merge this one tomorrow (and will check if removing `isidentifier()` improves a bit performances). I want to keep the first commit message as it indicates that `re` is now removed from the global namespace (it should never have been accessed from outside but we never know; maybe someone has been patching that attribute for whatever reason in their test suite).