@picnixz 2025-01-14
As a follow-up, I can also improve the import time of `pickletools` once I've merged this one (reason is that `pickletools` imports `re` and `pickle` so if I just change `pickletools` now, without this PR, then we won't see any improvements at all).