@picnixz 2025-01-14
@vstinner I plan to merge this one with the following commit message:

```text
Importing `pickle` is now roughly 25% faster.

Importing the `re` module is no longer needed and
thus is no more implicitly exposed as `pickle.re`.
```

and following title:

```text
Improve import time of the `pickle` module.
```